requirements.txt
openai-agents
python-dotenv
asyncio
streamlit